chore: refactor commands
- improve extensibility
- remove clutter
- improve organisation
