+++ regex/parser/include_with_definition_test.go
@@ -5,8 +5,12 @@ package parser
 
 import (
 	"bytes"
+	"fmt"
 	"io"
+	"io/fs"
 	"os"
+	"path"
+	"path/filepath"
 	"strings"
 	"testing"
 
@@ -19,11 +23,11 @@ import (
 
 type parserIncludeWithDefinitions struct {
 	suite.Suite
-	ctx         *processors.Context
-	reader      io.Reader
-	tempDir     string
+	rootDir     string
 	includeDir  string
 	includeFile []*os.File
+	ctx         *processors.Context
+	reader      io.Reader
 }
 
 func TestRunParserIncludeWithDefinitionsTestSuite(t *testing.T) {
@@ -64,4 +68,36 @@ func (s *parserIncludeWithDefinitions) TestParser_DanglingDefinitions() {
 	s.Equal(expected.String(), actual.String())
 
 	s.Contains(out.String(), "no match found for definition: {{hallo}}")
+}
+
+func (s *parserIncludeWithDefinitions) SetupSuite() {
+	s.rootDir = s.T().TempDir()
+	s.includeDir = path.Join(s.rootDir, "regex-assembly", "include")
+	err := os.MkdirAll(s.includeDir, fs.ModePerm)
+	s.Require().NoError(err)
+
+	rootContext := context.New(filepath.Dir(s.rootDir), "toolchain.yaml")
+	s.ctx = processors.NewContext(rootContext)
+	for i := 0; i < 4; i++ {
+		file, err := os.Create(path.Join(s.rootDir, fmt.Sprintf("multi-definitions-%d.ra", i)))
+		s.Require().NoError(err, "couldn't create %s file", file.Name())
+		if i == 0 {
+			// Only the initial include goes to the reader
+			s.reader = strings.NewReader(fmt.Sprintf(
+				"##!> include %s\n"+
+					"##!> define this-is-a-text [a-zA-J]+8\n"+
+					"##!> define this-is-another-text [0-9](pine|apple)\n"+
+					"{{this-is-a-text}} to see if definitions work.\n"+
+					"Second text for {{this-is-another-text}}.\n", file.Name()))
+		}
+		s.includeFile = append(s.includeFile, file)
+		// Write to file i-1
+		if i > 0 {
+			_, err := fmt.Fprintf(s.includeFile[i-1],
+				"##!> include %s\n"+
+					"This is comment %d.\n"+
+					"{{this-is-a-text}} to see if definitions work when included\n", file.Name(), i)
+			s.Require().NoError(err, "writing temp include file failed")
+		}
+	}
 }
